refactor(gpr): move GPR simplification into compression.py + run it in compress_model
PR #69 review: fold gpr_simplify.py into compression.py (delete the standalone
file) and have compress_model run simplify_model_gprs at the end of a
propagate_gpr pass. Now ANY caller of compress_model -- not just the SD pipeline
-- gets leaf-minimized GPR rules; standalone compression benefits too. Internal
helpers are prefixed _gpr_* to avoid any name overlap in compression.py.

Monotone/boolean-equivalent, so designs are unchanged; only the extend_model_gpr
gadget shrinks. In the pipeline simplify now runs in compress#1 AND again after
reduce_gpr, but it is cheap and idempotent. Design-identity green: test_04/05/07/
08/10/11/12 (241 tests). Updates the stale "simplification deferred to reduce_gpr"
notes to point at simplify_model_gprs.
